API-1488; fix comments
